requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
azure.storage.blob
anyio==4.1.0
azure-common==1.1.28
azure-core==1.29.6
azure-search-documents==11.4.0
azure-storage-blob==12.20.0
charset-normalizer==3.3.2
requests==2.31.0
openai==1.6.1




